+++ src/commands/index.ts
@@ -2,10 +2,10 @@ import { ForcecodeCommand, CancelCommand, FCCancellationToken } from './forcecod
 export { getApexTestResults } from '../services';
 import { compile, CompileMenu, ForceCompile } from './compile';
 import { DeployPackage, createPackageXML, deployFiles } from './deploy';
-import { diff, DiffMenu } from './diff';
+import { DiffMenu } from './diff';
 import { Open, ShowFileOptions } from './open';
 import { packageBuilder, PackageBuilder, getMembers, getFolderContents } from './packageBuilder';
-import { retrieve, RetrieveBundle, Refresh, ToolingType } from './retrieve';
+import { retrieve, RetrieveBundle, Refresh, ToolingType, getAnyNameFromUri } from './retrieve';
 import { getAuraDefTypeFromDocument, saveAura } from './saveAura';
 import { saveApex } from './saveApex';
 import { saveLWC } from './saveLWC';
@@ -37,11 +37,11 @@ import {
   Login,
   RemoveConfig,
   ChangeCoverageDecoration,
+  DeleteFile,
 } from './fcCommands';
 
 export {
   compile,
-  diff,
   packageBuilder,
   retrieve,
   saveApex,
@@ -58,6 +58,7 @@ export {
   createPackageXML,
   deployFiles,
   ToolingType,
+  getAnyNameFromUri,
 };
 
 export const fcCommands: ForcecodeCommand[] = [
@@ -104,4 +105,5 @@ export const fcCommands: ForcecodeCommand[] = [
   new Login(),
   new RemoveConfig(),
   new CancelCommand(),
+  new DeleteFile(),
 ];